fixed class name change for nav bar links, organized subfolders and components for jobs show page
