--- src/Components/Job/JobsShow.scss
@@ -1,220 +1,0 @@
-.jobShow {
-  display: grid;
-  grid-template-rows: 25vh 8vh 48vh 0.2vh 8vh;
-  row-gap: 0.5vh;
-  justify-items: left;
-  height: 92vh;
-  width: 100%;
-  color: var(--init-charcoal);
-  // padding-bottom: 2vh;
-
-  &_header {
-    width: 98%;
-    height: 100%;
-    justify-self: center;
-    grid-template-rows: 40% 15% 20% 15% 5%;
-    padding-top: 2vh;
-
-    &_logo {
-      height: 80px;
-      max-width: 100%;
-      grid-row: 1;
-      padding: 8px;
-      border-radius: 20px;
-    }
-
-    &_company {
-      grid-row: 2;
-      font-size: 18px;
-      margin-top: 1vh;
-    }
-    &_title {
-      grid-row: 3;
-    }
-    &_location {
-      grid-row: 4;
-      display: flex;
-      column-gap: 4vw;
-      align-items: center;
-
-      &_remote {
-        padding: 2px 5px;
-        border-radius: 10px;
-      }
-    }
-    & hr {
-      grid-row: 5;
-      border: 1px solid var(--init-darkmint);
-      width: 90%;
-    }
-  }
-
-  .skills-component {
-    grid-row: 2;
-    align-items: center;
-    justify-self: center;
-    border-radius: 20px;
-    background-color: white;
-    width: 86%;
-    height: 50%;
-    align-self: center;
-    margin: 0.6vh 0;
-    border: 2px solid var(--init-mint);
-    & svg {
-      height: 1.9em;
-      width: 1.9em;
-    }
-  }
-
-  .sliderButton {
-    &_container {
-      justify-self: center;
-      margin-left: 3vw;
-    }
-
-    &_buttons {
-      height: 5vh;
-
-      &_button {
-        height: 5vh;
-      }
-    }
-    &_slider {
-      height: 5vh;
-    }
-  }
-
-  .jobDetailSelected,
-  .jobShow_header_location_remote {
-    background-color: var(--init-darkmint);
-    color: var(--init-gold);
-    font-weight: 600;
-  }
-
-  &_jobDetails {
-    grid-row: 3;
-    width: 96%;
-    display: grid;
-    grid-template-rows: 5vh auto;
-    row-gap: 2vh;
-
-    &_overview {
-      overflow-y: scroll;
-      width: 92%;
-      padding: 1vh 0;
-      min-height: fit-content;
-      max-height: 52vh;
-      margin-left: 6vw;
-      font-size: 17px;
-
-      &_description {
-        display: grid;
-
-        &_disclaimer {
-          font-size: 16px;
-          position: relative;
-          bottom: 0.5vh;
-          font-weight: 600;
-          padding-bottom: 0.5vh;
-        }
-      }
-
-      &_tasks {
-        margin-top: 2vh;
-        padding-inline-start: 20px;
-        list-style-position: outside;
-        display: flex;
-        flex-direction: column;
-        row-gap: 2vh;
-        width: 88%;
-      }
-    }
-  }
-
-  &_divider {
-    width: 90%;
-    border: 1px solid var(--init-darkmint);
-    justify-self: center;
-  }
-
-  &_actionButtons {
-    display: grid;
-    grid-template-columns: 0.45fr 0.55fr;
-    width: 90%;
-    justify-self: center;
-    justify-items: center;
-    align-items: center;
-
-    & > * {
-      all: unset;
-      font-size: 30px;
-      color: var(--init-blue);
-      font-weight: 600;
-    }
-    &_apply,
-    &_applied,
-    &_notSignedIn {
-      grid-column: span 2;
-    }
-    &_notSignedIn {
-      display: flex;
-      width: 100%;
-      font-size: 24px;
-      justify-content: space-evenly;
-    }
-  }
-}
-
-.hide {
-  display: none;
-}
-
-/* DARK MODE */
-.dark .jobShow {
-  &_header {
-    color: var(--init-background);
-    &_logo {
-      background-color: var(--init-background);
-    }
-  }
-  &_jobDetails {
-    &_buttons {
-      color: var(--init-background);
-    }
-    &_overview {
-      color: var(--init-background);
-    }
-  }
-  &_actionButtons {
-    & > *,
-    &_notSignedIn a {
-      color: var(--init-gold);
-    }
-  }
-
-  .skills-component {
-    background-color: var(--init-background);
-  }
-
-  .jobDetailSelected,
-  .jobShow_header_location_remote {
-    background-color: var(--init-gold);
-    color: var(--init-dark);
-  }
-}
-
-
-@media (min-width: 450px) {
-  .jobShow {
-    .sliderButton {
-      &_container {
-        margin-left: 0vw;
-  }
-    }
-    &_jobDetails {
-      &_overview {
-        margin-left: 2.5vw;
-      }
-    }
-}
-}
