+++ src/Components/Job/JobsShowPage/DesktopJobsShowPage/DesktopJobsShow.scss
@@ -1,0 +1,3 @@
+.desktopJobShow {
+    height: 100vh;
+}
