+++ src/Components/Job/JobsShowPage/JobsShow.js
@@ -1,0 +1,74 @@
+import { useState, useEffect } from "react";
+// import { useNavigate, Link } from "react-router-dom";
+import { useWindowSizeProvider } from "../../../Providers/WindowSizeProvider";
+import { useContextProvider } from "../../../Providers/Provider";
+import { useNavProvider } from "../../../Providers/NavProvider";
+import { useJobProvider } from "../../../Providers/JobProvider";
+// import { useRecruiterProvider } from "../../../Providers/RecruiterProvider";
+import SkillsComponent from "../SkillsComponent";
+import JobsShowHeader from "./JobsShowHeader/JobsShowHeader";
+import JobsShowDetails from "./JobsShowDetails/JobsShowDetails";
+import JobsShowActionButtons from "./JobsShowActionButtons/JobsShowActionButtons";
+import "./JobsShow.scss";
+
+function JobsShow() {
+  const {isDesktopView} = useWindowSizeProvider()
+  const { API, axios, userID, isRecruiterAcc, isSignedIn, setLoading, loading } =
+    useContextProvider();
+  const { setAppHeader } = useNavProvider();
+  const { jobID, jobDetails, jobSkills } =
+    useJobProvider();
+  // const { editAccess } = useRecruiterProvider()
+  // const navigate = useNavigate();
+  const [applied, setApplied] = useState(false);
+  const [jobDetailsToggle, setJobDetailsToggle] = useState("description");
+
+
+  useEffect(() => {
+    setAppHeader("Job Details")
+    if(userID && jobID){
+      axios.get(`${API}/user-jobs/${userID}/${jobID}`)
+      .then(({data}) => 
+      data.date_applied ? 
+      setApplied(data.date_applied) : 
+      null)
+      .catch(err => console.log(err))
+    }
+  }, []);
+
+  useEffect(() => setLoading(false), [jobDetails])
+
+  return (
+    !loading && (
+      <div className="jobShow">
+        {
+          !isDesktopView ?
+          <JobsShowHeader jobDetails={jobDetails} /> :
+          <span>desktop header</span>
+        }
+       
+
+        <SkillsComponent skillsArr={jobSkills} justList={true} />
+
+        {
+          !isDesktopView ? 
+          <JobsShowDetails
+          jobDetails={jobDetails}
+          jobDetailsToggle={jobDetailsToggle}
+          setJobDetailsToggle={setJobDetailsToggle} /> :
+          <span>desktop</span>
+        }
+
+       
+        <hr className="jobShow_divider" />
+        
+        <JobsShowActionButtons 
+        applied={applied}
+        setApplied={setApplied} />
+  
+      </div>
+    )
+  );
+}
+
+export default JobsShow;
