+++ src/Components/Job/JobsShowPage/JobsShowActionButtons/JobsShowActionButtons.scss
@@ -1,0 +1,28 @@
+.jobShow {
+    &_actionButtons {
+        display: grid;
+        grid-template-columns: 0.45fr 0.55fr;
+        width: 90%;
+        justify-self: center;
+        justify-items: center;
+        align-items: center;
+    
+        & > * {
+          all: unset;
+          font-size: 30px;
+          color: var(--init-blue);
+          font-weight: 600;
+        }
+        &_apply,
+        &_applied,
+        &_notSignedIn {
+          grid-column: span 2;
+        }
+        &_notSignedIn {
+          display: flex;
+          width: 100%;
+          font-size: 24px;
+          justify-content: space-evenly;
+        }
+      }
+}
